deleted unnecessary files, updated hero text
