--- src/components/LandingComparison/landing-comparison.module.scss
@@ -1,238 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.section {
-  width: 100%;
-  background-color: var(--lit-black);
-  background: linear-gradient(
-    180deg,
-    var(--lit-black) 85%,
-    var(--lit-gray-900) 100%
-  );
-  color: var(--lit-gray-300);
-}
-
-.wrapper {
-  max-width: 72rem;
-  margin: 0 auto;
-  padding: 5rem 1rem;
-
-  @include sm {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-  }
-
-  @include lg {
-    padding: 8rem 4rem;
-  }
-}
-
-.header {
-  max-width: 56rem;
-  text-align: center;
-  margin: 0 auto;
-}
-
-.badge {
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 1rem;
-
-  @include lg {
-    margin-bottom: 1.5rem;
-  }
-
-  &__copy {
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    color: var(--lit-orange-light);
-    background-color: var(--lit-orange-light);
-    background-image: linear-gradient(
-      to bottom,
-      var(--lit-orange-light) 50%,
-      var(--lit-orange)
-    );
-    background-size: 100%;
-    -webkit-background-clip: text;
-    -moz-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    -moz-text-fill-color: transparent;
-
-    @include lg {
-      font-size: 1rem;
-      line-height: 1.5rem;
-    }
-  }
-}
-
-.title {
-  font-size: 1.875rem;
-  line-height: 2.25rem;
-  font-weight: bold;
-  color: var(--lit-gray-50);
-  background-color: var(--lit-gray-50);
-  background-image: linear-gradient(
-    to bottom,
-    var(--lit-gray-50),
-    var(--lit-gray-300)
-  );
-  background-size: 100%;
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-
-  @include xs {
-    font-size: 2.25rem;
-    line-height: 2.5rem;
-  }
-
-  @include lg {
-    font-size: 3rem;
-    line-height: 1;
-  }
-
-  &__break {
-    display: inline;
-
-    @include md {
-      display: block;
-    }
-  }
-}
-
-.subtitle {
-  font-size: 1rem;
-  line-height: 1.5rem;
-  font-weight: 300;
-  margin-top: 0.75rem;
-
-  @include lg {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    margin-top: 1rem;
-  }
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(12, minmax(0, 1fr));
-  gap: 1.5rem;
-  margin-top: 1.5rem;
-
-  @include md {
-    margin-top: 3rem;
-  }
-}
-
-.comparison {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  flex: 1;
-  padding: 1.5rem 1rem;
-  gap: 2rem;
-  grid-column: span 12;
-
-  @include sm {
-    padding-top: 2rem;
-    padding-bottom: 2rem;
-  }
-
-  @include md {
-    grid-column: 1 / span 6;
-  }
-
-  @include lg {
-    grid-column: 2 / span 5;
-    gap: 3rem;
-  }
-
-  &__title {
-    font-size: 1rem;
-    line-height: 1.5rem;
-    color: var(--lit-gray-500);
-
-    @include lg {
-      font-size: 1.125rem;
-      line-height: 1.75rem;
-    }
-  }
-
-  &__img {
-    width: 100%;
-    max-width: 20rem;
-    margin: auto;
-  }
-
-  &__badges {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 0.25rem;
-
-    @include xs {
-      gap: 0.5rem;
-    }
-  }
-
-  &__badge {
-    font-size: 0.75rem;
-    line-height: normal;
-    font-weight: 300;
-    padding: 0.375rem 0.75rem 0.25rem;
-    background-color: var(--lit-gray-900);
-    color: var(--lit-gray-400);
-    border: 1px solid transparent;
-    border-radius: 999px;
-    text-align: center;
-
-    @include lg {
-      font-size: 0.875rem;
-    }
-  }
-}
-
-.comparison--lit {
-  border-top: 1px solid var(--lit-gray-800);
-  grid-column: span 12;
-
-  @include md {
-    border-top: 0;
-    border-left: 1px solid var(--lit-gray-800);
-    grid-column: 7 / span 6;
-  }
-
-  @include lg {
-    grid-column: 7 / span 5;
-  }
-}
-
-.comparison--lit .comparison__title {
-  color: var(--lit-orange-light);
-  background-color: var(--lit-orange-light);
-  background-image: linear-gradient(
-    to bottom,
-    var(--lit-orange-light) 50%,
-    var(--lit-orange)
-  );
-  background-size: 100%;
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-}
-
-.comparison--lit .comparison__badge {
-  background-color: var(--lit-gray-700);
-  background-image: linear-gradient(
-    45deg,
-    var(--lit-gray-700),
-    var(--lit-gray-800)
-  );
-  color: var(--lit-gray-100);
-}
-
-.comparison--lit .comparison__img {
-  max-width: 15rem;
-}
