--- src/components/LandingCta/landing-cta.module.scss
@@ -1,261 +1,0 @@
-@import 'src/styles/mixins.scss';
-
-.section {
-  width: 100%;
-  background-color: var(--lit-white);
-  position: relative;
-}
-
-.wrapper {
-  max-width: 72rem;
-  margin: 0 auto;
-  padding: 5rem 1rem;
-
-  @include sm {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-  }
-
-  @include lg {
-    padding: 8rem 4rem;
-  }
-}
-
-.header {
-  text-align: center;
-}
-
-.title {
-  font-size: 1.875rem;
-  line-height: 2.25rem;
-  font-weight: bold;
-  color: var(--lit-gray-900);
-
-  @include xs {
-    font-size: 2.25rem;
-    line-height: 2.5rem;
-  }
-
-  @include sm {
-    font-size: 3rem;
-    line-height: 1;
-  }
-}
-
-.emphasis {
-  color: var(--lit-pink);
-  background-color: var(--lit-pink);
-  background-image: linear-gradient(
-    to bottom,
-    var(--lit-orange) 50%,
-    var(--lit-pink)
-  );
-  background-size: 100%;
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-}
-
-.subtitle {
-  color: var(--lit-gray-600);
-  font-size: 1.125rem;
-  line-height: 1.75rem;
-  font-weight: 300;
-  margin-top: 0.75rem;
-  margin-bottom: 1rem;
-
-  @include sm {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    margin-top: 1rem;
-    margin-bottom: 1.5rem;
-  }
-}
-
-.cta-btn {
-  display: inline-flex;
-  justify-content: center;
-  background-color: var(--lit-gray-900);
-  color: var(--lit-gray-50);
-  border: 1px solid var(--lit-gray-900);
-  border-radius: 999px;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  padding: 0.625rem 1rem 0.5rem;
-  transition: background-color 0.2s ease, border-color 0.2s ease;
-
-  &:hover {
-    background-color: var(--lit-orange);
-    border-color: var(--lit-orange);
-  }
-
-  @include sm {
-    font-size: 1rem;
-    line-height: 1.5rem;
-  }
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: auto auto;
-  gap: 1rem;
-  margin-top: 4rem;
-
-  @include md {
-    grid-template-columns: 1fr 1fr;
-  }
-}
-
-.card {
-  border-radius: 0.5rem;
-
-  &__img {
-    height: 8rem;
-    object-fit: cover;
-
-    @include sm {
-      height: 10rem;
-    }
-
-    @include md {
-      height: auto;
-    }
-  }
-
-  &__content {
-    padding: 1.5rem;
-
-    @include sm {
-      padding: 2rem;
-    }
-  }
-
-  &__title {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-
-    @include sm {
-      font-size: 1.5rem;
-      line-height: 2rem;
-    }
-
-    @include lg {
-      font-size: 1.875rem;
-      line-height: 2.25rem;
-    }
-  }
-
-  &__description {
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    margin-bottom: 1rem;
-
-    @include lg {
-      font-size: 1rem;
-      line-height: 1.5rem;
-    }
-  }
-
-  &__link {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.375rem;
-    border-radius: 999px;
-    border: 1px solid transparent;
-    font-size: 0.75rem;
-    line-height: 1rem;
-    padding: 0.5rem 0.875rem 0.375rem;
-    transition: background-color 0.2s ease, border-color 0.2s ease,
-      color 0.2s ease;
-
-    @include lg {
-      font-size: 0.875rem;
-      line-height: 1.25rem;
-      padding: 0.625rem 1rem 0.5rem;
-    }
-
-    &__icon {
-      width: 1rem;
-      height: auto;
-      margin-top: -2px;
-    }
-  }
-}
-
-.card--full {
-  grid-column: 1 / -1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 18rem;
-
-  @include sm {
-    min-height: 20rem;
-  }
-}
-
-.card--full .card__content {
-  max-width: 32rem;
-}
-
-.card--primary {
-  background-color: var(--lit-orange);
-}
-
-.card--primary .card__title {
-  color: var(--lit-white);
-  text-align: center;
-  margin-bottom: 0;
-}
-
-.card--dark {
-  background-color: var(--lit-gray-900);
-}
-
-.card--dark .card__title {
-  color: var(--lit-gray-50);
-}
-
-.card--dark .card__description {
-  color: var(--lit-gray-400);
-}
-
-.card--dark .card__link {
-  background-color: var(--lit-gray-900);
-  border-color: var(--lit-gray-700);
-  color: var(--lit-gray-200);
-
-  &:hover {
-    background-color: var(--lit-gray-800);
-    border-color: var(--lit-gray-600);
-    color: var(--lit-white);
-  }
-}
-
-.card--light {
-  background-color: var(--lit-gray-50);
-}
-
-.card--light .card__title {
-  color: var(--lit-gray-900);
-}
-
-.card--light .card__description {
-  color: var(--lit-gray-600);
-}
-
-.card--light .card__link {
-  background-color: var(--lit-gray-50);
-  border-color: var(--lit-gray-600);
-  color: var(--lit-gray-600);
-
-  &:hover {
-    background-color: var(--lit-gray-100);
-    color: var(--lit-gray-900);
-  }
-}
