--- src/components/LandingProducts/landing-products.module.scss
@@ -1,86 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.section {
-  width: 100%;
-  background-color: var(--lit-black);
-  background: linear-gradient(
-    180deg,
-    var(--lit-black) 85%,
-    var(--lit-gray-900) 100%
-  );
-  color: var(--lit-gray-300);
-}
-
-.wrapper {
-  max-width: 72rem;
-  margin: 0 auto;
-  padding: 5rem 1rem;
-
-  @include sm {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-  }
-
-  @include lg {
-    padding: 8rem 4rem;
-  }
-}
-
-.header {
-  max-width: 56rem;
-  text-align: center;
-  margin: 0 auto;
-}
-
-.title {
-  font-size: 1.875rem;
-  line-height: 2.25rem;
-  font-weight: bold;
-  color: var(--lit-white);
-  background-color: var(--lit-white);
-  background-image: linear-gradient(
-    to bottom,
-    var(--lit-white),
-    var(--lit-gray-300)
-  );
-  background-size: 100%;
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-
-  @include xs {
-    font-size: 2.25rem;
-    line-height: 2.5rem;
-  }
-
-  @include lg {
-    font-size: 3rem;
-    line-height: 1;
-  }
-
-  &__break {
-    display: inline;
-
-    @include md {
-      display: block;
-    }
-  }
-}
-
-.subtitle {
-  font-size: 1rem;
-  line-height: 1.5rem;
-  font-weight: 300;
-  margin-top: 0.75rem;
-
-  @include lg {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    margin-top: 1rem;
-  }
-}
-
-.divider {
-  border-top: 1px solid var(--lit-gray-800);
-}
