--- src/components/SubscribeForm/subscribe-form.module.scss
@@ -1,133 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.form {
-  margin-top: 1rem;
-
-  @include sm {
-    margin-top: 2rem;
-    padding: 0 1rem;
-  }
-
-  &__wrapper {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: stretch;
-    gap: 0.5rem;
-
-    @include sm {
-      flex-wrap: nowrap;
-    }
-  }
-
-  &__input {
-    width: 100%;
-    background-color: var(--lit-white);
-    color: var(--lit-gray-900);
-    padding: 0.625rem 1rem 0.5rem;
-    border: 2px solid transparent;
-    border-radius: 999px;
-    outline: none;
-
-    &:focus {
-      border-color: var(--lit-gray-900);
-    }
-
-    &:disabled {
-      background-color: var(--lit-gray-100);
-      cursor: none;
-    }
-
-    &::placeholder {
-      color: var(--lit-gray-500);
-    }
-  }
-
-  &__submit {
-    position: relative;
-    cursor: pointer;
-    display: inline-flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    gap: 0.375rem;
-    border-radius: 999px;
-    border: 2px solid transparent;
-    background-color: var(--lit-gray-900);
-    color: var(--lit-gray-50);
-    padding: 0.625rem 1rem 0.5rem;
-    transition: color 0.2s ease;
-
-    @include sm {
-      width: auto;
-      justify-content: center;
-    }
-
-    &:hover {
-      background-color: var(--lit-gray-800);
-    }
-
-    &:disabled {
-      background-color: var(--lit-gray-700);
-      cursor: none;
-    }
-
-    &--loading {
-      color: transparent;
-    }
-
-    &--loading::after {
-      content: '';
-      position: absolute;
-      width: 16px;
-      height: 16px;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      margin: auto;
-      border: 4px solid transparent;
-      border-top-color: var(--lit-white);
-      border-radius: 50%;
-      animation: spinner 1s ease infinite;
-    }
-
-    &__icon {
-      width: 1.5rem;
-      height: auto;
-      margin-top: -2px;
-    }
-  }
-
-  &__helper {
-    color: var(--lit-black);
-  }
-}
-
-@keyframes spinner {
-  from {
-    transform: rotate(0turn);
-  }
-
-  to {
-    transform: rotate(1turn);
-  }
-}
-
-.form__input:-webkit-autofill,
-.form__input:-webkit-autofill:focus {
-  transition: background-color 600000s 0s, color 600000s 0s;
-}
-.form__input[data-autocompleted] {
-  background-color: transparent !important;
-}
-
-.alert {
-  margin-top: 2rem;
-  width: 100%;
-  background-color: var(--lit-gray-900);
-  color: var(--lit-white);
-  padding: 0.625rem 1rem 0.5rem;
-  border: 2px solid var(--lit-gray-900);
-  border-radius: 999px;
-  text-align: center;
-}
